agent.py
from google.adk.agents.llm_agent import Agent

agent = Agent(
    name="learning_plan_agent",
    description="Creates weekly structured learning plans from beginner to advanced",
    instruction="""
    You are an expert learning mentor.

    The user will provide:
    1. Topic they want to learn
    2. Weekly hours they can spend

    Create a structured weekly learning roadmap from beginner to advanced.

    Output format:
    Week 1:
    Goal:
    Topics:
    Practical Task:

    Week 2:
    Goal:
    Topics:
    Practical Task:

    Continue progressively from basics to advanced.

    Keep the response concise and easy to follow.
    """
)