requirements.txt
fastapi
uvicorn
google-adk
google-genai
pydantic